WebGL::ClampPointers::convertArgumentToSmartStruct refactored
Previously there used to be two almost identical functions, differing
only at the type of the last argument and a few points. This change
embeds the differences of Instruction and BasicBlock to template
specialized struct LocationKind.

It is arguable that this can make the code more difficult to read, but
on the other hand it eliminates duplication.