Added some more unimplemented API calls to be able to finish connection between dependence analysis and address space info manager.
Basically what must be done is to keep track of every replacement to be able to ask from address space info manager what was the original value for this value and what is the current instruction corresponding this old one. In thery replace all uses could help except that we are have base limiting values in hand which will be ultimately deleted from SSA. This is pretty analoguous with the old way of keeping track of replaced function arguments.